requirements.txt
nltk
numpy
scikit-learn
PyPDF2
python-docx
